Переименовал папку, исправил pom.xml
