+++ src/main/java/ru/dnlkk/ratingusbackend/mapper/AnnouncementMapper.java
@@ -5,9 +5,9 @@ import org.mapstruct.Mapping;
 import org.mapstruct.Named;
 import org.mapstruct.factory.Mappers;
 import ru.dnlkk.ratingusbackend.api.model.AnnouncementDto;
-import ru.dnlkk.ratingusbackend.model.Announcement;
-import ru.dnlkk.ratingusbackend.model.Class;
-import ru.dnlkk.ratingusbackend.model.User;
+import ru.dnlkk.ratingusbackend.dtos.Announcement;
+import ru.dnlkk.ratingusbackend.dtos.Class;
+import ru.dnlkk.ratingusbackend.dtos.User;
 
 import java.util.ArrayList;
 import java.util.List;
