+++ src/main/java/ru/dnlkk/ratingusbackend/repository/AnnouncementRepository.java
@@ -2,9 +2,8 @@ package ru.dnlkk.ratingusbackend.repository;
 
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import ru.dnlkk.ratingusbackend.model.Announcement;
+import ru.dnlkk.ratingusbackend.dtos.Announcement;
 
 import java.util.List;
